requirements.txt
python-dateutil==2.8.1
psycopg2==2.8.6
numpy==1.9.3
gdal==3.0.2
rasterio==1.1.8
pandas==1.2.0
google-cloud-bigquery==2.6.1
six==1.14.0
pytz==2020.5
requests==2.25.1
